Fixed image issue
In line 292 onwards you had ....
                prize = PhotoImage(file="gold_med.gif")
                prize = "gold\n(${})".format(5 * stakes_multiplier)
Which means that your your 'prize' variable (which should have held the image) was being updated to hold the value.  You should have had...
                prize = PhotoImage(file="gold_med.gif")
                prize_list = "gold\n(${})".format(5 * stakes_multiplier)

Note the prize_list in the last line of the code above.  You have the same issue for each part of the if / elif / else block.  It's fixed in the code I'm proposing.