Errors Fixed (mostly)
Your code was very close to working.  Here are some notes on the changes I'd suggest to get it to do what you want it to do.  This is a partial fix but it should enable you to make progress and then continue to debug to get things working as expected.

On line 462, you are trying to disable a help_button but your button name is help_btn, which is why the program crashes.

Thing is, right now the button is being disabled when the help dialogue box is CLOSED and that is not what you want.  You want the button to be disabled when the help box is opened so that it is not possible to spawn multiple boxes.  To do this, change your code on *roughly* line 453 so the button is configured to 'NORMAL' rather than 'DISABLED'.

Your stats are not quite working because your code has the initial balance as a string rather than an integer.  To work this out, I printed the 'game_stats' list and noticed that the first item in the list, the starting balance, was in speech marks.

To fix this, in line 220, I added the line:
starting_balance = int(starting_balance)

I worked out that was the problem by going to where you printed the starting balance and asking it to print 'staring_balance*2', it output 2020 instead of 40 which helped to identify the fact that the starting balance was a string rather than an integer.